extracted Lanes class from monster-file DvcsGraph.cs to separate file with partial class
